+++ litebird_sim/bandpass_template_module.py
