+++ litebird_sim/pointings.py
@@ -2,7 +2,7 @@ import astropy.time
 import numpy as np
 import numpy.typing as npt
 
-from .hwp.hwp import HWP
+from .hwp import HWP
 from .scanning import (
     all_compute_pointing_and_orientation,
     RotQuaternion,
